refactor(notify): mount the emitter tests from their own file
The new batch and controller emitters pushed agent_task_notify.rs to 1657
lines, past the audit's 1500-line god_file threshold. There is no baseline
entry for this file, so that would have been a brand-new finding -- and adding
one would trip the baseline ratchet from #11777, which pins the count in both
directions.

Relocate the test module behind #[path], following cook.rs -> cook_tests.rs.
Pure relocation: 958 + 706 lines, all 28 tests still pass.
